refactor(acceptance): extract scenario 26 helpers into sibling file (warren-8f17)
